+++ lib/commands/slicemachine/commands/sync.js
@@ -1,7 +1,6 @@
 import consola from 'consola';
 import ora from 'ora';
 import inquirer from 'inquirer';
-import path from 'path';
 
 import Communication from '../../../services/communication';
 import { Files, Folders, Objects } from '../utils';
@@ -12,7 +11,6 @@ import Sentry from '../../../services/sentry';
 
 import { getOrFail as getSmFile } from '../methods/sm';
 import Libraries from '../common/libraries';
-import { exception } from 'console';
 
 
 // writes the custom types in the filesystem
