+++ lib/commands/slicemachine/utils/index.js
@@ -1,0 +1,4 @@
+export { default as Objects } from './objects';
+export { default as Files } from './files';
+export { default as Folders } from './folders';
+export { default as ziputils } from './ziputils';
